+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - `REMIND` statement
 - `MOVE` entries to a different time
 
+### Changed
+- Output is now colored
+
 ## 0.1.0 - 2021-12-20
 
 ### Added